# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AI Go Chi Starter API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Liveness probe
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '504':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
  /readyz:
    get:
      summary: Readiness probe
      responses:
        '200':
          description: Ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '503':
          description: Not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '504':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
  /v1/examples:
    get:
      summary: List examples
      responses:
        '200':
          description: Example list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '503':
          description: Service draining or not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '504':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
    post:
      summary: Create example
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExampleRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '503':
          description: Service draining or not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '504':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
  /v1/examples/{id}:
    get:
      summary: Get example by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Example detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '503':
          description: Service draining or not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '504':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
components:
  schemas:
    Envelope:
      type: object
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        data:
          oneOf:
            - $ref: '#/components/schemas/Example'
            - type: array
              items:
                $ref: '#/components/schemas/Example'
            - type: object
              additionalProperties: true
            - type: 'null'
    CreateExampleRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    Example:
      type: object
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time